Some improvements
This was originally meant to add CLI to this program, but then I saw some flaws with the gui on windows, so these got fixed. Also the working directory, I was wondering where everything was getting saved, so that was changed to a suitable directory in the directory where the jar is executed.

Enjoy!